requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-pinecone
pinecone
chromadb
pypdf
python-dotenv
